componentização da sidebar e content
